refactor: update types and interfaces for TableFilterList and related components
